Implement compose() feature DSL in SlackExtension
This moves compose support to a new DSL configuration instead.

```kotlin
slack {
  android {
    features {
      compose()
    }
  }
}
```

Currently there is no extra configuration beyond the above in the DSL, but we could add more in the future. This also adds a new `defaultComposeAndroidBundleAlias` property that can be used to identify a default `libs.bundle.*` alias to include in any projects that enable compose. This is, in this case, just focused on adding _core_ compose deps (runtime, ui, etc).

One implementation note is that due to AGP using regular kotlin properties still, we have to make enabling compose a synchronous action that directly sets them on the underlying AGP property under the hood. If they ever convert them into lazy gradle properties, we could look at chaining these instead.