Introduce monkey patch for MergeFileTask (#224)
After some investigation, I believe this task is currently breaking
remote caching for lint because it has non-deterministic ordering of
file merges, resulting in the merged file's content to be different.
This introduces a monkey patch of the task (that is - we'll put it on
the classpath first so that AGP uses our implementation) that has a
system property guard to enable sorting of these files first. This
allows us to test this theory out. I've also created a standalone
monkeypatch artifact as a toe hold for any future ones we want to add
too.

Note the task is copied identically from AGP sources with the exception
of the extra block for filtering files.

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->