Consolidate common lint logic to LintTasks (#281)
This is similar in structure to #280 but for lints. It creates new
`ciLint` and a root `globalCiLint` tasks in the project to ease running
things on CI.

This also extracts a lot of lint logic from the bloated
`StandardProjectConfigurations` file to this standalone bit of code.

One notable change is that now we run lint for all variants in a project
unless otherwise specified. This seems like the right thing to do as we
should also be running lint on internal/debug/etc sources too.

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->