Add extra skippy diagnostics to SlackTools
This improves our Skippy diagnostics by writing a diagnostics file to the `build/logs` dir at the end of the build with a list of all the skipped tasks. This makes it a bit easier to debug which tasks are being explicitly skipped. Also cleaned up some of the StandardProjectConfigurations setup along the way.