Add extra skippy diagnostics to SlackTools (#338)
This improves our Skippy diagnostics by writing a diagnostics file to
the `build/logs` dir at the end of the build with a list of all the
skipped tasks. This makes it a bit easier to debug which tasks are being
explicitly skipped. Also cleaned up some of the
StandardProjectConfigurations setup along the way.

Resolves #340

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->