Consolidate compose compiler configuration
Testing new versions of the compose compiler with new versions of kotlin is a massive PITA. The compose compiler version checks are tedious and that's multiplied by compose multiplatform. This consolidates all of this logic into a shared place while also offering a control to force use of the androidx compose compiler in multiplatform projects when it's raced ahead.