Persistent State of Components with Changelog Parsing  (#457)
This is a (preliminary) PR for the data storage of the latest entry seen
with the change log parsing for the skate plugin.

`ChangelogParser`:
The primary responsibility of `ChangelogParser` is to parse and present
the changes from a changelog file. The parsed information is stored with
`PersistentStateComponent`. This component maintains a State object that
stores the last read date of the changelog.

Helper functions/Regex created:
1. `LOCAL_DATE_REGEX`: Matches dates in "yyyy-mm-dd" format.
2. `VERSION_PATTERN_REGEX`: Matches version patterns in "\d+.\d+.\d+"
format.
3. `String.isLocalDate `and `String.startsNewBlock` are extension
functions that use the above regex patterns to validate if a string can
be parsed as a date and if a new block starts, respectively.

The `readFile` function takes the changelog string as an argument and
processes it to extract new changes and update the `lastReadDate` state.

I plan to delete `ChangelogParser` which was the code previous to adding
the states.


<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->

---------

Co-authored-by: Zac Sweers <pandanomic@gmail.com>