Fix authenticator plugin (#509)
Continuing from #500. Also had to remove the okio dependency as it
appears to conflict with dependencies coming from the intellij plugin
now, so using standard java to do it. Tested this with studio giraffe
(the bugged version) and it works now

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->