Add tracing to Skate plugin (#617)
Add the first set of metrics to be tracked for Skate plugin, including

* Whats New Panel:
      * When the What's new panel opened / closed
* Feature Flag Annotator:
      * When user clicked on Houston link from feature flag annotation
 * Project Generator:
* When the terminal opened and execute command to spin up ProjectGen UI

Noted that for the WhatsNewPanel, I added a custom
`WhatsNewToolWindowListener` to subscribe to the event for the window
opened/closed. I followed the set up described
[here](https://plugins.jetbrains.com/docs/intellij/plugin-listeners.html#defining-project-level-listeners)
to set up the listener


Testing:
Example traces sent:
https://analytics.tinyspeck.com/v3/explore/1000000001135498


  

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->