Granularize skippy configs for future work (#678)
This is a part one toward exposing granular skippy configurations per
tool. Currently it's still implemented in the project computer the same
as before and just merges them all together, but will continue that work
in subsequent PRs.

One thing I'm curious for thoughts on in this PR is if this approach of
having the AffectedProjectsComputer makes sense, or if it might be
better to keep it implemented the same as before and run _multiple_
computer runs in parallel instead (where each task corresponds to the
tool it's computing for). Then at the end we could aggregate their
computed affected projects files and merge them. I'm leaning toward the
latter but keen for thoughts!

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->