Streamline kotlin configuration and JDK setup (#883)
This is a sort of two part cleanup PR
1. Streamline kotlin configuration with properties for various configs
and refactored logic. This also includes a new control for a language
version override.
2. Streamline jdk/jvm target sourcing to just SlackVersions and use lazy
providers for all of them

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->