Revert Jewel Update  (#947)
Reverts PR #937 as it was causing issues with the Skate plugin. It needs
to be upgraded in the future along with migrating to IntelliJ Platform
Gradle Plugin 2.x.
<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->