Skate AI Initial Chat User Text Field and Window (#948)
This is a continuation of #927, but because of the Jewel update in #937,
I ended up running into some hiccups with transitioning the project and
migrating to a different version of intellij gradle platform plugin. We
reverted it back in #947 and this PR reflects the changes that are
compatible with the old Jewel version (which is back on main).

I have included the functionality for user input, where send button
clears text, enter clears text, and shift + enter adds new lines. Users
can also add multiple lines. I also added a function where the send
button is faded and unselectable until there is text input. In future
PRs, I will add the chat conversation as well, and this is just a start.


https://github.com/user-attachments/assets/94971a43-476e-4d67-924c-e012c633bc3c

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->