Chat Bubble and "Conversation"  (#950)
In this PR, I added a Chat bubble that will print back the user input. I
also set up a function "call api" that will call the DevXP API in the
future, but currently sends back the user input as well. The
conversation field is also scrollable (with no scroll bar at the
moment).

I am also worried about how there is not much padding once the text
input becomes multiple lines (like in example 2 of the video), and was
wondering if I could get advice on how to add it, because adding padding
modifiers seems to only change the box and not the inner text, and the
height parameter is what creates the initial space in the text field.


https://github.com/user-attachments/assets/50c6bf01-a77a-4b4d-950e-a849ad70ef16

EDIT: here is an update based on feedback. Now, the user's messages are
on the right and response is on the left.


https://github.com/user-attachments/assets/62170207-3697-41e5-ac0e-3ff24b4cb1fd

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/slack-gradle-plugin/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->