Disable AI Chat Window for New Skate Release (#1023)
Just commented out the tool window for now because it's not ready yet.
So we can do a new Skate release that's compatible with K2.
<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/foundry/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->