Add isLoading State to Chat (#1018)
Adding a state for if it's loading once the API is called, so that a
loading animation can be added while waiting for the chat response.

Below is an example of the animation. I enabled it so it's always
loading just to show the animation (since the API call hasn't been
merged yet)

https://github.com/user-attachments/assets/7abb3041-7aa0-4117-94e6-710a6b0cb800


<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/foundry/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->