Improve Skate publishing
Now we get tags and automatic changelog management. This tunes the publish command to now take either a specific version or an automatic version via `--major/--minor/--patch` argument.

`./publish_skate --major` or `./publish_skate 0.8.1`.

We should now start adding to the "Unreleased" section when we make changes, similar to what we do with the top-level CHANGELOG.md for foundry changes.