Initial pass at ModuleTopographyTask (#1070)
This PR introduces two new tasks to Foundry.

1. `ModuleTopographyTask` (name: `moduleTopography`). This task reads
various information about a module's build features and dumps them to a
JSON file. At the moment these features hard coded directly in source,
but in the future we could look at trying to make this a more
configurable JSON format that could be configurable per-repo. The idea
is that this offers a way to get insight into what build features a
module has in a way that other tools can then read from. For example -
mod score could read this rather than its current, non-reusable
implementation.
2. `ValidateModuleTopographyTask` (name: `validateModuleTopography`).
This is a sister task to the first one, reading its outputs and then
just validating that each _declared_ feature is actually _used_. If it's
declared but not used, it fails validation and can optionally be
auto-fixed.

Next after this PR:
- Make mod score read this
- Explore consumer-side JSON config files
- Add more detectable features
- Explore sourcing from DAGP outputs to determine other features like
resources usage or android API/aar dep usage.

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/foundry/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->