Make robolectric sdk downloads less hardcoded (#1107)
This reworks how we handle robolectric jar downloads by allowing the
consumer side to potentially force a newer version and load SDKs from
the newer robolectric SDK. Still not super ideal as it requires putting
robolectric on the buildscript classpath, but it's an impl dep at least
and hopefully we can iterate on it in the future (maybe making the
coordinate downloads a direct download rather than use the gradle
resolver?)

I think eventually it would be helpful to spin this out to a separate,
reusable plugin as we're definitely not the only ones with this issue.

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/foundry/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->