Add validation task infrastructure + java version matcher
This introduces a new concept of `validateFoundryProject` as a lifecycle task that can depend on a few common validations. The first one I'm adding here is a new one to validate that a `.java_version` file (if present) matches the `jdk` version defined in a version catalog, which we wanna support to make some other tooling more automatic like renovate and github actions.