Add validation task infrastructure + java/kotlin version matcher (#1247)
This introduces a new concept of `validateFoundryProject` as a lifecycle
task that can depend on a few common validations. The first one I'm
adding here is a new one to validate that a `.java_version` file (if
present) matches the `jdk` version defined in a version catalog, which
we wanna support to make some other tooling more automatic like renovate
and github actions.

Example of the .java_version file
https://github.com/square/moshi/pull/1941/files

<!--
  ⬆ Put your description above this! ⬆

  Please be descriptive and detailed.
  
Please read our [Contributing
Guidelines](https://github.com/tinyspeck/foundry/blob/main/.github/CONTRIBUTING.md)
and [Code of Conduct](https://slackhq.github.io/code-of-conduct).

Don't worry about deleting this, it's not visible in the PR!
-->