Add Noisy Skate Banner to move Android Studio Installation (#1346)
This PR adds a Skate noisy banner to nudge developers to move their
Android Studio installation to `desired/path` if it's not already
detected to be installed in that path.

**How this PR was Tested**
1. I tested this out side by side with Android Studio Installations in
my original path, and in the new desired path.
2. I built a zip file of Skate with `./gradlew
:platforms:intellij:skate:buildPlugin` and installed this zip into the
respective Android Studios. The one in the original path had the warning
message, while the one in `desired/path` did not. I also have the path
posted in the warning message for more visibility.

I also pre-populated the settings below:
- Installation location pattern (glob): `desired/path`
- More info URL: `slack message url`