Remove AI bot code (#1400)
Since the project was paused, this feature is no longer in use, so I’d
like to remove it. We’ll still have a record of it here if we need to
revisit it in the future. I also moved `painterResource` to a `utils`
folder since it can be reused by other Skate features.